[fee-payer] Add multi-agent / fee payer support
This adds 3 new APIs, prepRawTransaction, signRawTransaction, and
signAndSubmitRawTransaction.  Prep will create a multiagent or fee
payer transaction accordingly, unsigned with information from the
wallet on sequence number and chain id.  Then all parties can sign
the transaction with signRawTransaction, and the last party can
sign with signAndSubmitRawTransaction to sign and send it off as the
sender.